feat: complete interopt cpp

+++ LowLevelCalculator8/src/include/utils.h
@@ -1,13 +1,11 @@
 #ifndef __UTILS_H
 #define __UTILS_H
 
-#include <iostream>
-
 int absolute(int x);
 int addition(int x, int y);
 int substraction(int x, int y);
 int multiplication(int x, int y);
-float division(int x, int y);
+int division(int x, int y);
 int power(int x, int y);
 
 #endif
